Tests/LibWeb/Text/expected/WebAudio/BaseAudioContext-decodeAudioData.txt

channels: 1
length: 8
sampleRate: 8000
channel0.length: 8
sample[0]: -1.000000
sample[1]: -0.500000
sample[2]: 0.000000
sample[3]: 0.500000
sample[4]: 0.999969
sample[5]: 0.250000
sample[6]: -0.250000
sample[7]: 0.000031
